ref(tracemetrics): Extract state keeping from state changes in provider
This generalizes the state changes so I can change the way we store state from
flushing to the URL to keeping local state. This will be necessary for Alerts
where Alerts should only store the final equation. I only need to parse the
equation and locally render the subcomponents.
